additional warning flags and fixes various bugs: * fixes overlapping communication with KDDecomposition - querynonblocking was broken * akward handling of simulation in Homogeneous and Planar class * fixes checks for AVX2 in VectorizedCellProcessor, which led to no fastreciprocal sqareroots being used
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@5514 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/utils/AlignedArray.h
@@ -113,7 +113,7 @@ public:
 	virtual void prefetch(int hint = 1, int n = -1) const {
 		mardyn_assert(n >= -2);
 
-		int endPrefetch;
+		size_t endPrefetch;
 		const int stride = _round_up(1);
 
 		switch(n) {
@@ -130,7 +130,7 @@ public:
 			endPrefetch = n;
 		}
 
-		for (int i = 0; i < endPrefetch; i+= stride) {
+		for (size_t i = 0; i < endPrefetch; i+= stride) {
 			const T & val = _vec[i];
 			const T * valP = &val;
 #if defined(__MIC__)
